+++ android/app/src/main/java/com/daynest/android/core/di/NetworkApiDiModule.kt
@@ -1,0 +1,47 @@
+package com.daynest.android.core.di
+
+import com.daynest.android.data.calendar.CalendarApi
+import com.daynest.android.data.medication.MedicationApi
+import com.daynest.android.data.settings.SettingsApi
+import com.daynest.android.data.templates.TemplatesApi
+import com.daynest.android.data.today.PlannedItemApi
+import com.daynest.android.data.today.TodayActionsApi
+import com.daynest.android.data.today.TodayApi
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NetworkApiDiModule {
+    @Provides
+    @Singleton
+    fun provideTodayApi(retrofit: Retrofit): TodayApi = retrofit.create(TodayApi::class.java)
+
+    @Provides
+    @Singleton
+    fun provideTodayActionsApi(retrofit: Retrofit): TodayActionsApi = retrofit.create(TodayActionsApi::class.java)
+
+    @Provides
+    @Singleton
+    fun providePlannedItemApi(retrofit: Retrofit): PlannedItemApi = retrofit.create(PlannedItemApi::class.java)
+
+    @Provides
+    @Singleton
+    fun provideCalendarApi(retrofit: Retrofit): CalendarApi = retrofit.create(CalendarApi::class.java)
+
+    @Provides
+    @Singleton
+    fun provideMedicationApi(retrofit: Retrofit): MedicationApi = retrofit.create(MedicationApi::class.java)
+
+    @Provides
+    @Singleton
+    fun provideTemplatesApi(retrofit: Retrofit): TemplatesApi = retrofit.create(TemplatesApi::class.java)
+
+    @Provides
+    @Singleton
+    fun provideSettingsApi(retrofit: Retrofit): SettingsApi = retrofit.create(SettingsApi::class.java)
+}
