+++ android/app/src/main/java/com/daynest/android/core/di/NetworkDiModule.kt
@@ -7,13 +7,6 @@ import com.daynest.android.core.network.CertificatePinnerProvider
 import com.daynest.android.core.network.DynamicBaseUrlInterceptor
 import com.daynest.android.core.network.JsonSerializer
 import com.daynest.android.core.network.TokenAuthenticator
-import com.daynest.android.data.calendar.CalendarApi
-import com.daynest.android.data.medication.MedicationApi
-import com.daynest.android.data.settings.SettingsApi
-import com.daynest.android.data.templates.TemplatesApi
-import com.daynest.android.data.today.PlannedItemApi
-import com.daynest.android.data.today.TodayActionsApi
-import com.daynest.android.data.today.TodayApi
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -89,32 +82,4 @@ object NetworkDiModule {
             .client(okHttpClient)
             .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
             .build()
-
-    @Provides
-    @Singleton
-    fun provideTodayApi(retrofit: Retrofit): TodayApi = retrofit.create(TodayApi::class.java)
-
-    @Provides
-    @Singleton
-    fun provideTodayActionsApi(retrofit: Retrofit): TodayActionsApi = retrofit.create(TodayActionsApi::class.java)
-
-    @Provides
-    @Singleton
-    fun providePlannedItemApi(retrofit: Retrofit): PlannedItemApi = retrofit.create(PlannedItemApi::class.java)
-
-    @Provides
-    @Singleton
-    fun provideCalendarApi(retrofit: Retrofit): CalendarApi = retrofit.create(CalendarApi::class.java)
-
-    @Provides
-    @Singleton
-    fun provideMedicationApi(retrofit: Retrofit): MedicationApi = retrofit.create(MedicationApi::class.java)
-
-    @Provides
-    @Singleton
-    fun provideTemplatesApi(retrofit: Retrofit): TemplatesApi = retrofit.create(TemplatesApi::class.java)
-
-    @Provides
-    @Singleton
-    fun provideSettingsApi(retrofit: Retrofit): SettingsApi = retrofit.create(SettingsApi::class.java)
 }
